Auto-updating electronic beer pong table code
This code is meant to run an electronic beer pong table equipped with
photoresistors under the cups to detect whether or not there is a cup
present on the table. Based on this reading, the LCD screens are
updated to show the current score and the LEDs are changed accordingly.